+++ src/Plugin/WebformHandler/GetOrganizedWebformHandler.php
@@ -36,6 +36,8 @@ class GetOrganizedWebformHandler extends WebformHandlerBase {
    * @var \Drupal\Core\Logger\LoggerChannelInterface
    */
   protected LoggerChannelInterface $submissionLogger;
+  private const string ADDITIONAL = 'additional';
+  private const string STATES = 'states';
 
   /**
    * Constructs a GetOrganizedWebformHandler object.
@@ -71,6 +73,19 @@ class GetOrganizedWebformHandler extends WebformHandlerBase {
       $container->get('webform_submission.conditions_validator'),
       $container->get('webform.token_manager')
     );
+  }
+
+  /**
+   * {@inheritdoc}
+   *
+   * @phpstan-return array<string, mixed>
+   */
+  public function defaultConfiguration() {
+    return [
+      self::ADDITIONAL => [
+        self::STATES => [WebformSubmissionInterface::STATE_COMPLETED],
+      ],
+    ];
   }
 
   /**
@@ -211,6 +226,31 @@ class GetOrganizedWebformHandler extends WebformHandlerBase {
       ],
     ];
 
+    // Additional.
+    // Lifted from EmailWebformHandler::buildConfigurationForm().
+    $resultsDisabled = (bool) $this->getWebform()->getSetting('results_disabled');
+    $form[self::ADDITIONAL] = [
+      '#type' => 'fieldset',
+      '#title' => $this->t('Additional settings'),
+    ];
+    // Settings: States.
+    $states = (array) ($this->configuration[self::ADDITIONAL][self::STATES] ?? NULL);
+    $form[self::ADDITIONAL][self::STATES] = [
+      '#type' => 'checkboxes',
+      '#title' => $this->t('Run handler when …'),
+      '#options' => [
+        WebformSubmissionInterface::STATE_DRAFT_CREATED => $this->t('<b>draft is created</b>.'),
+        WebformSubmissionInterface::STATE_DRAFT_UPDATED => $this->t('<b>draft is updated</b>.'),
+        WebformSubmissionInterface::STATE_CONVERTED => $this->t('anonymous <b>submission is converted</b> to authenticated.'),
+        WebformSubmissionInterface::STATE_COMPLETED => $this->t('<b>submission is completed</b>.'),
+        WebformSubmissionInterface::STATE_UPDATED => $this->t('<b>submission is updated</b>.'),
+        WebformSubmissionInterface::STATE_DELETED => $this->t('<b>submission is deleted</b>.'),
+        WebformSubmissionInterface::STATE_LOCKED => $this->t('<b>submission is locked</b>.'),
+      ],
+      '#access' => !$resultsDisabled,
+      '#default_value' => $resultsDisabled ? [WebformSubmissionInterface::STATE_COMPLETED] : $states,
+    ];
+
     return $this->setSettingsParents($form);
   }
 
@@ -229,6 +269,11 @@ class GetOrganizedWebformHandler extends WebformHandlerBase {
     $this->configuration['choose_archiving_method']['cpr_value_element'] = $form_state->getValue('choose_archiving_method')['cpr_value_element'];
     $this->configuration['choose_archiving_method']['cpr_name_element'] = $form_state->getValue('choose_archiving_method')['cpr_name_element'];
     $this->configuration['choose_archiving_method']['sub_case_title'] = $form_state->getValue('choose_archiving_method')['sub_case_title'];
+
+    $additional = $form_state->getValue(self::ADDITIONAL);
+    // Clean up states.
+    $additional[self::STATES] = array_values(array_filter($additional[self::STATES]));
+    $this->configuration[self::ADDITIONAL] = $additional;
   }
 
   /**
@@ -264,8 +309,9 @@ class GetOrganizedWebformHandler extends WebformHandlerBase {
    * {@inheritdoc}
    */
   public function postSave(WebformSubmissionInterface $webform_submission, $update = TRUE): void {
-    // Ignore anything that is not completed.
-    if (!$webform_submission->isCompleted()) {
+    $submissionState = $webform_submission->getWebform()->getSetting('results_disabled') ? WebformSubmissionInterface::STATE_COMPLETED : $webform_submission->getState();
+    $enabledStates = (array) ($this->configuration[self::ADDITIONAL][self::STATES] ?? NULL);
+    if (!in_array($submissionState, $enabledStates)) {
       return;
     }